requirements.txt
aiogram==3.*
deep-translator==1.11.*
google-genai==0.1.*